# jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Solution to combine municipality services under single API.",
    "version" : "0.1.0",
    "title" : "Kunta API"
  },
  "host" : "demo.kuntaapi.fi",
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "paths" : {
    "/electronicServiceChannels" : {
      "get" : {
        "tags" : [ "ServiceChannels", "ElectronicServiceChannels" ],
        "summary" : "Lists electronic service channels",
        "description" : "Lists electronic service channels",
        "operationId" : "listElectronicServiceChannels",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns array of electronic service channels",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ElectronicServiceChannel"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/electronicServiceChannels/{electronicServiceChannelId}" : {
      "get" : {
        "tags" : [ "ServiceChannels", "ElectronicServiceChannels" ],
        "summary" : "Finds an electronic service channel by id",
        "description" : "Finds an electronic service channel by id",
        "operationId" : "findElectronicServiceChannel",
        "parameters" : [ {
          "name" : "electronicServiceChannelId",
          "in" : "path",
          "description" : "electronicChannel id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns an electronic service channel by id",
            "schema" : {
              "$ref" : "#/definitions/ElectronicServiceChannel"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
    "/phoneServiceChannels" : {
      "get" : {
        "tags" : [ "ServiceChannels", "PhoneServiceChannels" ],
        "summary" : "Lists phone service channels",
        "description" : "Lists phone service channels",
        "operationId" : "listPhoneServiceChannels",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns array of phone service channels",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/PhoneServiceChannel"
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/phoneServiceChannels/{phoneServiceChannelId}" : {
      "get" : {
        "tags" : [ "ServiceChannels", "PhoneServiceChannels" ],
        "summary" : "Finds a phone service channel by id",
        "description" : "Finds a phone service channel by id",
        "operationId" : "findPhoneServiceChannel",
        "parameters" : [ {
          "name" : "phoneServiceChannelId",
          "in" : "path",
          "description" : "Phone service channel id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a phone service channel",
            "schema" : {
              "$ref" : "#/definitions/PhoneServiceChannel"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
    "/services/{serviceId}" : {
      "get" : {
        "tags" : [ "Services" ],
        "summary" : "Finds a service by id",
        "description" : "Finds service",
        "operationId" : "findService",
        "parameters" : [ {
          "name" : "serviceId",
          "in" : "path",
          "description" : "Service id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a service",
            "schema" : {
              "$ref" : "#/definitions/Service"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Services" ],
        "summary" : "Updates service",
        "description" : "Updates service",
        "operationId" : "updateService",
        "parameters" : [ {
          "name" : "serviceId",
          "in" : "path",
          "description" : "Organization service id",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Payload",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Service"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns an updated service",
            "schema" : {
              "$ref" : "#/definitions/Service"
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "schema" : {
              "$ref" : "#/definitions/BadRequest"
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "schema" : {
              "$ref" : "#/definitions/Forbidden"
            }
          },
          "404" : {
            "description" : "Resource was not found from the server",
            "schema" : {
              "$ref" : "#/definitions/NotFound"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "basicAuth" : {
      "type" : "basic"
    }
  },
  "definitions" : {
...
    "Service" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "type" : {
          "type" : "string"
        },
        "statutoryDescriptionId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "chargeType" : {
          "type" : "string"
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "areaType" : {
          "type" : "string",
          "description" : "Area type (WholeCountry, WholeCountryExceptAlandIslands, AreaType)."
        },
        "areas" : {
          "type" : "array",
          "description" : "List of service areas.",
          "items" : {
            "$ref" : "#/definitions/Area"
          }
        },
        "descriptions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "languages" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "serviceClasses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "ontologyTerms" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "targetGroups" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "lifeEvents" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "industrialClasses" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OntologyItem"
          }
        },
        "legislation" : {
          "type" : "array",
          "description" : "List of laws related to the service.",
          "items" : {
            "$ref" : "#/definitions/Law"
          }
        },
        "keywords" : {
          "type" : "array",
          "description" : "List of localized service keywords.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "requirements" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "publishingStatus" : {
          "type" : "string",
          "description" : "Publishing status. Possible values are: Draft, Published, Deleted, Modified or OldPublished."
        },
        "organizations" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServiceOrganization"
          }
        },
        "electronicServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "phoneServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "printableFormServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "serviceLocationServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "webPageServiceChannelIds" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      }
    },
    "ServiceOrganization" : {
      "type" : "object",
      "properties" : {
        "additionalInformation" : {
          "type" : "array",
          "description" : "Localized list of additional information.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "organizationId" : {
          "type" : "string",
          "description" : "Organization identifier (organization related to the service). Required if role type is Responsible or if ProvisionType is SelfProduced."
        },
        "roleType" : {
          "type" : "string",
          "description" : "Role type, valid values Responsible or Producer."
        },
        "provisionType" : {
          "type" : "string",
          "description" : "Provision type, valid values SelfProduced, VoucherServices, PurchaseServices or Other. Required if RoleType value is Producer."
        },
        "webPages" : {
          "type" : "array",
          "description" : "List of web pages.",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        }
      },
      "description" : "Service organization"
    },
...
    "ElectronicServiceChannel" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid"
        },
        "organizationId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "descriptions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "signatureQuantity" : {
          "type" : "integer",
          "format" : "int32"
        },
        "requiresSignature" : {
          "type" : "boolean"
        },
        "supportPhones" : {
          "type" : "array",
          "description" : "List of support phone numbers for the service channel.",
          "items" : {
            "$ref" : "#/definitions/Phone"
          }
        },
        "supportEmails" : {
          "type" : "array",
          "description" : "List of support email addresses for the service channel.",
          "items" : {
            "$ref" : "#/definitions/Email"
          }
        },
        "requiresAuthentication" : {
          "type" : "boolean"
        },
        "urls" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "languages" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "attachments" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServiceChannelAttachment"
          }
        },
        "webPages" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        },
        "serviceHours" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServiceHour"
          }
        },
        "publishingStatus" : {
          "type" : "string"
        }
      }
    },
    "PhoneServiceChannel" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Identifier for the service channel."
        },
        "organizationId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "Organization identifier responsible for the channel."
        },
        "names" : {
          "type" : "array",
          "description" : "Localized list of service channel names.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "descriptions" : {
          "type" : "array",
          "description" : "List of localized service channel descriptions.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "phoneNumbers" : {
          "type" : "array",
          "description" : "List of phone numbers for the service channel.",
          "items" : {
            "$ref" : "#/definitions/Phone"
          }
        },
        "supportEmails" : {
          "type" : "array",
          "description" : "List of support email addresses for the service channel.",
          "items" : {
            "$ref" : "#/definitions/Email"
          }
        },
        "languages" : {
          "type" : "array",
          "description" : "List of languages the service channel is available in (two letter language code).",
          "items" : {
            "type" : "string"
          }
        },
        "webPages" : {
          "type" : "array",
          "description" : "List of service channel web pages.",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        },
        "serviceHours" : {
          "type" : "array",
          "description" : "List of service channel service hours.",
          "items" : {
            "$ref" : "#/definitions/ServiceHour"
          }
        },
        "publishingStatus" : {
          "type" : "string",
          "description" : "Service channel publishing status. Values: Draft, Published, Deleted, Modified or OldPublished."
        }
      }
    },
...
    "WebPage" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string",
          "description" : "Web page description."
        },
        "url" : {
          "type" : "string",
          "description" : "Web page url."
        },
        "language" : {
          "type" : "string",
          "description" : "Language code."
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "Email" : {
      "type" : "object",
      "properties" : {
        "value" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        }
      }
    },
    "Phone" : {
      "type" : "object",
      "properties" : {
        "additionalInformation" : {
          "type" : "string"
        },
        "serviceChargeType" : {
          "type" : "string"
        },
        "chargeDescription" : {
          "type" : "string"
        },
        "prefixNumber" : {
          "type" : "string"
        },
        "isFinnishServiceNumber" : {
          "type" : "boolean"
        },
        "number" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "ServiceChannelAttachment" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "url" : {
          "type" : "string"
        },
        "language" : {
          "type" : "string"
        }
      }
    },
...
    "ServiceHour" : {
      "type" : "object",
      "properties" : {
        "serviceHourType" : {
          "type" : "string"
        },
        "validFrom" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date time where from this entry is valid."
        },
        "validTo" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date time to this entry is valid."
        },
        "isClosed" : {
          "type" : "boolean",
          "description" : "Set to true to present a time between the valid from and to times as closed."
        },
        "validForNow" : {
          "type" : "boolean",
          "description" : "Set to true to present that this entry is valid for now."
        },
        "additionalInformation" : {
          "type" : "array",
          "description" : "Localized list of additional information.",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "openingHour" : {
          "type" : "array",
          "description" : "List of servicing hours (open and closes times).",
          "items" : {
            "$ref" : "#/definitions/DailyOpeningTime"
          }
        }
      }
    },
    "DailyOpeningTime" : {
      "type" : "object",
      "properties" : {
        "dayFrom" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Day index (0 == Sunday)"
        },
        "dayTo" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Day index (0 == Sunday)"
        },
        "from" : {
          "type" : "string",
          "description" : "Start time for example 10:00:00"
        },
        "to" : {
          "type" : "string",
          "description" : "End time for example 20:00:00"
        },
        "isExtra" : {
          "type" : "boolean",
          "description" : "Set to true to have extra time for a day. Enables to have open times like 10:00-14:00 and also on the same day 16:00-20:00. So the latter time is extra time."
        }
      }
    },
    "OntologyItem" : {
      "type" : "object",
      "properties" : {
        "system" : {
          "type" : "string"
        },
        "name" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "code" : {
          "type" : "string"
        },
        "ontologyType" : {
          "type" : "string"
        },
        "uri" : {
          "type" : "string"
        },
        "parentId" : {
          "type" : "string",
          "format" : "uuid"
        },
        "parentUri" : {
          "type" : "string"
        }
      }
    },
    "Law" : {
      "type" : "object",
      "properties" : {
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "webPages" : {
          "type" : "array",
          "description" : "List of localized web page urls.",
          "items" : {
            "$ref" : "#/definitions/WebPage"
          }
        }
      },
      "description" : "a Law"
    },
...
    "LocalizedValue" : {
      "type" : "object",
      "properties" : {
        "language" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        }
      }
    },
    "Municipality" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "string",
          "description" : "Municipality code (like 491 or 091)."
        },
        "names" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        }
      }
    },
    "Area" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string",
          "description" : "Type of the area (Municipality, Province, BusinessRegions, HospitalRegions)."
        },
        "code" : {
          "type" : "string",
          "description" : "Code of the area."
        },
        "name" : {
          "type" : "array",
          "description" : "Localized list of names for the area",
          "items" : {
            "$ref" : "#/definitions/LocalizedValue"
          }
        },
        "municipalities" : {
          "type" : "array",
          "description" : "List of municipalities including municipality code and a localized list of municipality names.",
          "items" : {
            "$ref" : "#/definitions/Municipality"
          }
        }
      },
      "description" : "Area"
    },
...
    "InternalServerError" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "Forbidden" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
    "BadRequest" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
...
    "NotFound" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "message" : {
          "type" : "string"
        }
      }
    }
  },
...
}